extract console to its own class
